OAuth2_test/requirements.txt
annotated-types
anyio
certifi
fastapi
h11
httpcore
httpx
idna
Jinja2
pydantic
pydantic_core
pydantic-settings
pytest
python-dotenv
sniffio
starlette
typing_extensions
uvicorn